Remove unused StdError
